+++ packages/react/src/components/cookieConsent/components/CookieSettingsPage.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { useCookieSettingsPage } from '../hooks/useCookieSettingsPage';
+
+export const CookieSettingsPage = () => {
+  const { settingsPageId } = useCookieSettingsPage();
+
+  return <div id={settingsPageId} />;
+};