+++ packages/react/src/components/cookieConsentCore/CookieSettingsPage.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import { useCookieSettingsPage } from './hooks/useCookieSettingsPage';
-
-export const CookieSettingsPage = () => {
-  const { settingsPageId } = useCookieSettingsPage();
-
-  return <div id={settingsPageId} />;
-};